refactor: 사용자(User) 도메인 포인트 자료형 변경, validation 로직 메서드 분리
- 포인트 자료형 변경 Integer -> BigDecimal
- 자료형 변경에 따른 User 도메인, 통합, E2E 테스트 코드 수정

+++ apps/commerce-api/src/main/java/com/loopers/application/user/UserFacade.java
@@ -8,6 +8,8 @@ import com.loopers.support.error.ErrorType;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @RequiredArgsConstructor
 @Component
 public class UserFacade {
@@ -31,7 +33,7 @@ public class UserFacade {
         return UserInfo.from(user);
     }
 
-    public Integer getUserPoint(String userId) {
+    public BigDecimal getUserPoint(String userId) {
         User findUser = userService.getUserPointByUserId(userId);
 
         if (findUser == null ) {
@@ -41,7 +43,7 @@ public class UserFacade {
         return findUser.getPoint();
     }
 
-    public UserInfo chargeUserPoint(String userId, Integer chargePoint) {
+    public UserInfo chargeUserPoint(String userId, BigDecimal chargePoint) {
         User findUser = userService.chargePointByUserId(userId, chargePoint);
 
         return UserInfo.from(findUser);
