+++ apps/commerce-api/src/main/java/com/loopers/application/user/UserInfo.java
@@ -3,7 +3,9 @@ package com.loopers.application.user;
 import com.loopers.domain.user.Gender;
 import com.loopers.domain.user.User;
 
-public record UserInfo(Long id, String userId, String email, String birthdate, Gender gender, Integer point) {
+import java.math.BigDecimal;
+
+public record UserInfo(Long id, String userId, String email, String birthdate, Gender gender, BigDecimal point) {
     public static UserInfo from(User user) {
         return new UserInfo(
                 user.getId(),
