+++ apps/commerce-api/src/main/java/com/loopers/interfaces/api/user/UserV1DTO.java
@@ -5,6 +5,8 @@ import com.loopers.domain.user.Gender;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 
+import java.math.BigDecimal;
+
 public class UserV1DTO {
 
     public record UserRequest(
@@ -16,10 +18,10 @@ public class UserV1DTO {
 
     public record UserPointRequest(
             @NotBlank(message = "사용자 ID는 필수값입니다.") String userId,
-            Integer chargePoint
+            BigDecimal chargePoint
     ) { }
 
-    public record UserResponse(Long id, String userId, String email, String birthdate, Gender gender, Integer point) {
+    public record UserResponse(Long id, String userId, String email, String birthdate, Gender gender, BigDecimal point) {
         public static UserResponse from(UserInfo userInfo) {
             return new UserResponse(
                     userInfo.id(),
@@ -32,8 +34,8 @@ public class UserV1DTO {
         }
     }
 
-    public record UserPointResponse(String userId, Integer point) {
-        public static UserPointResponse from(String userId, Integer point) {
+    public record UserPointResponse(String userId, BigDecimal point) {
+        public static UserPointResponse from(String userId, BigDecimal point) {
             return new UserPointResponse(
                     userId,
                     point
